Add arbitrary anchor to the Histogram class
 The Histogram class was slightly changed adding an arbitrary anchor 
as reference bin, which before was fixed to be 0. This is needed for 
a to-be-added method to shift the histogram in a rigid way by an 
aribtrary amount. Without an anchor which has also to be shifted, 
shifting the edges of the bins by an arbitrary amount but still using 
0 as anchor would be wrong.