Fix severe bug in reweighting of probability
distributions A vector was constructed with a known size, but 
push_back was used to fill it up resulting in twice the elements, 
half of which where zeroes. Being these numbers the bins height 
estimators for the error on the heights of the bins, having the 
double of the estimators set to zero and the half set to number 
fluctuating around the bins height was artificially increasing the 
errors.