Compress sample data files
For testing purposes, some quite large data files were added to the
repository. Compressing them saves more than 80% space and it is
worth, especially since cmake is able to uncompress the archive
setting up the build folder for compilation. Note that at this
commit, the history of the repository was rewritten to completely
remove the previously-scattered-around files. If you checkout an
older commit than this one you will probably not able to run the full
cmake setup, because cmake will try to copy not existing files. You
can e.g.run cmake before checking out back in history, so that you
have the files somewhere.