Fix compilation with Apple clang
The initializer_list members of the observables classes (Variance,
Skewness and Kurtosis) were declared as constexpr, which enters a kind
of dark corner of the modern standards (it is unclear which should be
the correct behaviour of the compiler). This was accepted by gcc 7.5 on
Ubuntu but refused by Apple clang version 11.0.3 and therefore we
decided to make them just const (the Mean by the way was already so and
therefore there was an inconsistency).