Let Quantities contain Mean Variance Skewness and Kurtosis
Although maybe this might not be strictly speaking necessary,
Quantities has now a member of each single quantity type. At the moment
these are public, but the next step is to make them private and
implement an access operator[] which will take a string and will return
a reference to the public EstimateAndError member of the single
quantities. This will structure access a bit and we should then be
ready to add to Quantities a constructor with a DataSample and a
Parameter to be then used on the dataAnalysis side.