Move tests for mean and use Mean object to calculate it
The mean and its error was calculated using the
calcMeanAndErrorOfDataSample function in dataAnalysisUtilities file and
this has to be changed, since in the end we want to delegate the
calculation to the Mean object. Tests where this function was used have
been moved and refactored in order to use the Mean object. Some further
work is now required, because it has been noticed that to pass around
to/in the single quantity classes the full Parameters object is not
needed since just the binning information (which binning strategy and
the number of bins/binsize) is needed (beyond the bool to know if the
mean is zero which is in any case stored in the quantity class). It is
probably worth to create a "binningParameters" object which has to be
then given together with a bool isMeanKnownToBeZero.