Extract constants into namespace instead of static members
This has been done for all static variables of the single quantity
classes. The clear advantage of having a namespace is that it reduces
interdependency of cmake libraries. On top, we transformed these
variables into template variables, with a fully specialization per
quantity. The minimum required standard is then C++17, though. At the
moment it is not possible to let the template not implemented to avoid
its usage with wrong types, but the default values should probably
trigger errors elsewhere, if used (e.g. we set the unsigned integers
to UINT_MAX and this physically is highly improbable to make sense, for
example as needed moment).