Avoid using jackknife for mean and variance evaluation
Although this was the initial idea, doing an unnecessary jackknife
analysis for mean and variance implies doing more algebra with
consequent larger rounding errors. Tests that were passing for men and
variance in dataAnalysisUtilities without using the jackknife were then
failing once the test were migrated to quantities and adapted in order
to use the Mean and Variance classes. This has been then adjusted
avoiding jackknife making the protected method to make the jackknife
calculation in AbstractQuantity virtual and overriding it in the Mean
and in the Variance classes (the parent virtual method implementation
will then be used for Skewness and Kurtosis).