Use MultipleDataSample in Quantities
An algorithm to use multiple estimators in the analysis of quantities
is still missing, but the code is then in a better shaipe to add
one. The new object MultipleDataSample has a to-be-implemented method
to evaluate the n-th moment using several estimators. This can be then
used in the single quantities calculation.