Extract QuantityAttributes object and use it where needed
This is the first step to be able to enable the analysis with multiple
estimates in the single quantity classes. In this way the
QuantityAttributes class has a second bool protected member that
contains infomation about the usage of multiple estimates. This is
analogous of isMeanZero. These two bool variables are now moving around
together in the QuantityAttributes struct from Parameters to the
Quantities instantiation.