Throw in general template quantities function definitions
This will make the case in which such functions are by accident called
easier to detect. A C++17 if-constexpr has been used to avoid a
false-positive warning while picking up the correct function for the
Mean due to use of std::is_same_v but not forcing it to be evaluated at
compile time.