Treat Variance as higher moments if multiple estimates used
With one single estimates per trajectory it has been already noticed in
the past that performing a Jackknife analysis, although theoretically
correct, would lead to rounding errors that can be avoided calculating
directly the value and the error. However, this is not possible any
more in the case when multiple estimates are used and, hence, it makes
sense in this case to use the general Jackknife analysis provided in
the QuantityAbstract class (used already by Skewness and Kurtosis).