Implement check on calculated quantity
When using multiple estimates it can happen that weird results are
obtained like a negative variance/kurtosis or nan for the
skewness. This is the case e.g. with artificial samples and/or with too
few estimates. Hence we implemented a check for these cases and, at the
moment, we throw an exception if anything like that happens.