Implement Jackknife functions when using multiple estimates
The Jackknife function must be different and in particular it must use
the expanded version of the higher central moments, since it would be
wrong to evaluate a central moment when using multiple
estimates. Please, note that some code duplication has been left on
purpose to add clarity to what is being done.