Make function to get unbiased moment a template function
This function was written in the case the moment per trajectory is
estimated out of multiple estimates, but in general the same machinery
is used on any bunch of moments estimators in the Reweighting and hence
the idea is to accept as parameter a std::vector<T> of any type that
can be manipulated in an algebraic sense. More specifically, T will be
either a realFloat or a DataSample. For the later, the pow function is
still to be provided.