Remove use of useMultipleEstimates bool where not needed
Now that it has been clarified that the usage of multiple estimates
belongs to raw data and that actually the bias has to be removed
evaluating the moments per trajectory, it follows that there is no need
to do anything related with the bias reduction after the reweighting
procedure. What it had been implemented is, strictly speaking, not
wrong and it consisted in basically evaluate unbiased second, third and
fourth moments ahead of reweighting with exclusively 4 estimates and
each estimate for the mean was reweighted separately and averaged out
after the reweighting to have the estimate for the first moment.
Correct but definitely confusing and unnecessarily specific. The idea
is to build unbiased moments ahead of reweighting in the
SimulationDataContainer class and then simply reweight them as in the
case of not using multiple estimates. Therefore, after having
reweighted them, it will make no difference whether multiple estimates
were used when it comes to combine moments to build up
Quantities. Hence, on the quantities side, all the functionality
dealing with Moments and MomentsEstimators does not need to care about
any useMultipleEstimate bool. This simplified the code pretty much. Now
the next step is to fix the Reweighting starting from the moments
calculation in the SimulationDataContainer class.