Adjust procedure to build unbias moments
The SimulationDataContainer class was only working with 4 as number of
multiple columns to build moments and the first moment was not
built. Instead, the columns were left unchanged for later
reweighting. This is now changed in a way that, no matter which column
are told to be used, these are used to build the demanded moments and
the bunch of columns are always replaced by as many as the demanded
moments.