Make reweighting correctly work with multiple estimates
The reweighting is now just letting the SimulationDataContainer prepare
moments and then it reweights them without any need to have several
estimates as result at the new points. This simplified the code a lot,
since no bunch of columns for the first moment has to be treated in a
different manner. Tests were also adjusted accordingly.