Add check of quantities to reweighted case and do it always
At first, the check was implemented in the multiple estimate case only,
but whether this is the case for moments obtained in the reweighting is
then difficult to be figured out and it would e.g. imply to let the
reweighting core machinery know about several estimate usage, which is
actually not any more the case. Since the check is anyhow cheap, we now
always do it.