Fix extraction binning number in case of multiple estimates
The binning procedure has to be done using a binsize that depends on
the moments to be used, due to the Jackknife analysis. In turn, hence,
the determination of this number depends on which expression for the
quantity is used, i.e. whether central or non-central moments are to be
used.